:sparkles: 광야 토큰 갱신, 반환 로직 재구성
