+++ src/main/kotlin/team/themoment/gsmNetworking/GsmNetworkingApplication.kt
@@ -2,7 +2,9 @@ package team.themoment.gsmNetworking
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.scheduling.annotation.EnableScheduling
 
+@EnableScheduling
 @SpringBootApplication
 class GsmNetworkingApplication
