+++ src/main/kotlin/team/themoment/gsmNetworking/domain/community/domain/GwangyaToken.kt
@@ -7,8 +7,11 @@ import org.springframework.data.redis.core.index.Indexed
 import java.time.LocalDateTime
 
 @RedisHash("gwangya_token")
-class GwangyaToken(
+data class GwangyaToken(
     @Id
+    val id: String,
+
+    @Indexed
     val gwangyaToken: String,
 
     @Indexed
